app/requirements.txt
Django==2.2.6
django-cors-headers==3.1.1
django-redis==4.10.0
djangorestframework==3.10.3
gunicorn==19.9.0
pytz==2019.3
redis==3.3.8
sqlparse==0.3.0
psycopg2==2.8.3
